Save wizard informations to database and deploy contract to testnet

+++ 8-Model/Company.cs
@@ -6,8 +6,8 @@ namespace Auctus.Model
     public class Company
     {
         public String Name { get; set; }
-        public Decimal BonusFee { get; set; }
-        public Decimal MaxBonusFee { get; set; }
+        public Double BonusFee { get; set; }
+        public Double MaxBonusFee { get; set; }
 
         public IEnumerable<VestingRules> VestingRules { get; set; }
     }
