+++ 8-Model/Employee.cs
@@ -5,8 +5,8 @@ namespace Auctus.Model
     public class Employee
     {
         public String Name { get; set; }
-        public Decimal Salary { get; set; }
-        public Decimal ContributionPercentage { get; set; }
+        public Double Salary { get; set; }
+        public Double ContributionPercentage { get; set; }
     }
 }
