+++ DomainObjects/Funds/PensionFundOption.cs
@@ -10,8 +10,8 @@ namespace Auctus.DomainObjects.Funds
     public class PensionFundOption
     {
         public String Address { get; set; }
-        public Decimal Fee { get; set; }
-        public Decimal LatePenalty { get; set; }
+        public Double Fee { get; set; }
+        public Double LatePenalty { get; set; }
         public Int32 PensionFundId { get; set; }
 
         public Wallet Wallet { get; set; }
